8166026: Refactor java/lang shell tests to java
